+++ includes/inc.php
@@ -31,8 +31,8 @@ $cryptoGlance = new CryptoGlance();
 $GLOBALS['cryptoGlance'] = $cryptoGlance;
 $settings = $cryptoGlance->getSettings();
 
-//// Current Build:
-define('CURRENT_VERSION', 'v2.1.0.13');
+// Current Build:
+define('CURRENT_VERSION', 'v2.1.0.14');
 
 // Misc function used throughout cryptoglance
 require_once('functions.php');