+++ index.php
@@ -61,13 +61,14 @@ require_once("includes/header.php");
         }
 
     /* PoolPicker */
-        if ($cryptoGlance->isConfigAvailable('poolpicker')) {
-            // Load specific JS for this panel
-            $jsArray[] = 'dashboard/PoolPicker';
-
-            $poolPicker = $cryptoGlance->getPoolPicker();
-            require_once("templates/panels/poolpicker.php");
-        }
+    /* This service has Shut Down */
+        // if ($cryptoGlance->isConfigAvailable('poolpicker')) {
+        //     // Load specific JS for this panel
+        //     $jsArray[] = 'dashboard/PoolPicker';
+        //
+        //     $poolPicker = $cryptoGlance->getPoolPicker();
+        //     require_once("templates/panels/poolpicker.php");
+        // }
 
     /* Wallets */
       if ($cryptoGlance->isConfigAvailable('wallets')) {
@@ -87,7 +88,7 @@ require_once("includes/header.php");
 
       require_once("templates/modals/add_rig.php");
       require_once("templates/modals/add_pool.php");
-      require_once("templates/modals/poolpicker.php");
+    //   require_once("templates/modals/poolpicker.php");
 ?>
    </div>
    <!-- /container -->